Moved AllToolsMenu UI components into their own files and group
